requirements.txt
matplotlib==3.9.3
numpy==1.23.5
opencv_python==4.10.0.84
Pillow==11.0.0
scikit_learn==1.6.0
seaborn==0.13.2
tensorflow==2.10.1
flask==3.1.0
streamlit==1.18.0

